Move polearm staff aura script into warrior spells
